requirements.txt
# Core Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.3
pillow==10.2.0

# Screen capture
mss==9.0.1

# Machine learning (using CPU version for Windows compatibility)
tensorflow-cpu>=2.13.0
# OR pytorch (if preferred)

# Data processing and analysis
pandas==2.1.4

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# GUI and input handling
pynput==1.7.6

# Configuration
pyyaml==6.0.1

# Additional utilities
scikit-learn==1.4.0
scipy==1.11.4
